More pythonic
Update the data to be lowercase before matching.

Drop useless dedup(). Drop needless intermediate step of chemical symbols. 

The fel=[] state is a little ugly. I'd have moved that to a generator function so that the main code is 
`fel = list(extract_chemical_elements(letters))`
and the `for c in letters` code has yields instead of `fel.append`